+++ cores/esp32/nefry/NefryWeb.cpp
@@ -55,7 +55,7 @@ String Nefry_Web::beginWeb(String _viwePage) {
 			"</div><div class=\"writemode\">"
 			"</div><ul>"
 			"<li><a href='/wifi_conf'>Setup WiFi</a>"
-			"<li><a href='/module_id'>Setup Module</a>"
+			"<li><a href='/config'>Data Store</a>"
 			"<li><a href='/web_update'>Web Sketch Download</a>"
 			"<li><a href='/update'>Upload Sketch</a>"
 			"<li><a href='/console'>Web Console</a>"
